+++ config/import_export.php
@@ -12,6 +12,7 @@ return [
     'tmp_path' => 'import-export/tmp',
     'queue' => 'import-export',
     'batch_size' => 500,
+    'max_batches_per_job' => 10000,
     'max_file_size' => 52428800,
     'retention_days' => 30,
     'error_cap_per_severity' => 1000,